feat(uncertainty)!: Use Tyche provider
BREAKING CHANGE: Replace dynamic pseudo-Sobol/Morris and local conformal/statistics contracts with Tyche-backed const-generic APIs.

Refs: backlog.md#kw-uq-062--integrate-tyche-uncertainty-ownership-major-arch--ready-review

+++ crates/kwavers-solver/src/inverse/pinn/ml/uncertainty_quantification/types.rs
@@ -10,9 +10,9 @@ pub struct PinnUncertaintyConfig {
     /// Ensemble size for deep ensembles.
     pub ensemble_size: usize,
     /// Conformal prediction alpha (1-confidence level).
-    pub conformal_alpha: f64,
+    pub conformal_alpha: f32,
     /// Variance threshold for reliability assessment.
-    pub variance_threshold: f64,
+    pub variance_threshold: f32,
 }
 
 /// Prediction with uncertainty bounds.
